+++ GST.h
@@ -1,16 +1,16 @@
 #pragma once
 //
 //    FILE: GST.h
-// VERSION: 0.1.0
+// VERSION: 0.1.1
 // PURPOSE: Arduino library for Gold Standard Test metrics
 //     URL: https://github.com/RobTillaart/GST
 //          https://en.wikipedia.org/wiki/Sensitivity_and_specificity
 //          https://en.wikipedia.org/wiki/Confusion_matrix
 //
-// formula's based upon wikipedia.
+// formula's based upon Wikipedia.
 
 
-#define GST_LIB_VERSION           (F("0.1.0"))
+#define GST_LIB_VERSION           (F("0.1.1"))
 
 
 class GST
@@ -18,12 +18,20 @@ class GST
 public:
   GST() {};
 
-  // These 4 need to be filled in.
-  void setTruePositive(float v)   { TP = v; P = TP + FN; };
-  void setTrueNegative(float v)   { TN = v; N = TN + FP; };
-  void setFalsePositive(float v)  { FP = v; N = TN + FP; };
-  void setFalseNegative(float v)  { FN = v; P = TP + FN; };
+  //  These four values of the matrix need to be set to get started.
+  void setTruePositive(float v = 0)   { TP = v; P = TP + FN; };
+  void setTrueNegative(float v = 0)   { TN = v; N = TN + FP; };
+  void setFalsePositive(float v = 0)  { FP = v; N = TN + FP; };
+  void setFalseNegative(float v = 0)  { FN = v; P = TP + FN; };
 
+  //  These are used for updating the test matrix
+  void addTruePositive(float v)   { TP += v; P = TP + FN; };
+  void addTrueNegative(float v)   { TN += v; N = TN + FP; };
+  void addFalsePositive(float v)  { FP += v; N = TN + FP; };
+  void addFalseNegative(float v)  { FN += v; P = TP + FN; };
+
+
+  //  Basic output functions
   float getTruePositive()   { return TP; };
   float getTrueNegative()   { return TN; };
   float getFalsePositive()  { return FP; };
@@ -39,42 +47,35 @@ public:
   float specificity()       { return TNR(); };
 
 
-
-  //  true positive rate
+  float truePositiveRate()  { return TPR(); };
   float TPR() { return TP / P; };
-  //  true negative rate
+  float trueNegativeRate()  { return TNR(); };
   float TNR() { return TN / N; };
-
-  //  false negative rate
-  float FNR() { return FN / (FN + TP); };
-  //  false positive rate
-  float FPR() { return FP / (FP + TN); };
-
+  float falseNegativeRate() { return FNR(); };
+  float FNR() { return FN / P; };
+  float falsePositiveRate() { return FPR(); };
+  float FPR() { return FP / N); };
 
 
-  //  positive predictive value
+  float positivePredictiveValue() { return PPV(); };
   float PPV() { return TP / (TP + FP); };
-  //  negative predictive value
+  float negativePredictiveValue() { return NPV(); };
   float NPV() { return TN / (TN + FN); };
-
-  //  false discovery rate
+  float falseDiscoveryRate() { return FDR(); };
   float FDR() { return FP / (FP + TP); };
-  //  false omission rate
+  float falseOmissionRate() { return FOR(); };
   float FOR() { return FN / (FN + TN); };
 
 
-
-  //  positive likelihood ratio
+  float positiveLikelihoodRatio() { return LRplus(); };
   float LRplus()  { return TPR() / FPR(); };
-  //  negative likelihood ratio
+  float negativeLikelihoodRatio() { return LRminus(); };
   float LRminus() { return FNR() / TNR(); };
-
 
 
   float prevalenceThreshold()  { return sqrt(FPR()) / (sqrt(TPR()) + sqrt(FPR())); };
   float threatScore()  { return TP / (TP + FN + FP); };
   float criticalSuccessIndex() { return threatScore(); };
-
 
 
   float prevalence()       { return P / (P + N); };
@@ -83,18 +84,19 @@ public:
   float F1Score()          { return (2 * TP)/(2 * TP + FP + FN); };
 
 
-
-  //  Matthews correlation coefficient
-  float MCC() { return (TP*TN-FP*FN)/sqrt((TP+FP)*(TP+FN)*(TN+FP)*(TN+FN)); };
+  float MatthewsCorrelationCoefficient() { return MMC(); };
   float phi() { return MCC(); };
-  //  Fowlkes–Mallows index
+  float MCC() { return (TP*TN - FP*FN)/sqrt((TP+FP) * (TP+FN) * (TN+FP) * (TN+FN)); };
+  float FowlkesMallowsIndex() { return FM(); };
   float FM() { return sqrt(PPV()*TPR()); };
-  //  Bookmaker informedness
+  float BookmakerInformedness() { return BM(); };
   float BM() { return TPR() + TNR() - 1; };
-  //  markedness
-  float MK() { return PPV() + NPV() - 1; };
+
+
+  float markedness() { return MK(); };
   float deltaP() { return MK(); };
-  //  diagnostic odds ratio
+  float MK() { return PPV() + NPV() - 1; };
+  float diagnosticOddsRatio() { return DOR(); };
   float DOR() { return LRplus() / LRminus(); };
 
 
@@ -106,6 +108,12 @@ private:
   float FP = 0;
   float FN = 0;
 };
+
+
+//  for now identical.
+class GoldenStandardTest::GST
+{
+}
 
 
 // -- END OF FILE --