+++ changelog.md
@@ -2,7 +2,17 @@
 # GST Changelog
 
 
-## 0.1.0 2022-02-25
+## 0.1.1  2022-06-08
+
+- add **addTruePositive()** etc functions.
+- add defaults for **setTruePositive(value = 0)** etc functions
+- add long descriptive names for the short functions.
+- add derived class GoldenStandardTest, for descriptive name
+- added some documentation
+
+
+## 0.1.0  2022-02-25
+
 - initial version
 -